Knowledge Base and Research Agent/Knowledge Base and Research Agent/pyproject.toml
[project]
name = "knowledge-base-and-research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chainlit>=2.7.2",
    "ddgs>=9.5.4",
    "numpy>=2.2.6",
    "openai-agents>=0.2.9",
    "pdfplumber>=0.11.7",
    "pymupdf>=1.26.4",
    "pypdf2>=3.0.1",
]
